centralise epoch resolution in standalone file to ease migration to v2
