+++ scripts/gas-refund-program/transactions-indexing/fetchRefundableTransactionsAllChains.ts
@@ -1,11 +1,12 @@
 import { assert } from 'ts-essentials';
-import { CHAIN_ID_MAINNET } from '../../../src/lib/constants';
-import { EpochInfo } from '../../../src/lib/epoch-info';
 import {
   GasRefundGenesisEpoch,
   GRP_SUPPORTED_CHAINS,
 } from '../../../src/lib/gas-refund';
-import { resolveEpochCalcTimeInterval } from '../common';
+import {
+  getCurrentEpoch,
+  resolveEpochCalcTimeInterval,
+} from '../epoch-helpers';
 import {
   getLatestEpochRefunded,
   merkleRootExists,
@@ -15,8 +16,6 @@ import { fetchPricingAndTransactions } from './fetchPricingAndTransactions';
 const logger = global.LOGGER('GRP::fetchRefundableTransactionsAllChains');
 
 export async function fetchRefundableTransactionsAllChains() {
-  const epochInfo = EpochInfo.getInstance(CHAIN_ID_MAINNET, true);
-
   return Promise.all(
     GRP_SUPPORTED_CHAINS.map(async chainId => {
       const lastEpochRefunded = await getLatestEpochRefunded(chainId);
@@ -30,11 +29,7 @@ export async function fetchRefundableTransactionsAllChains() {
         'cannot compute refund data for epoch < genesis_epoch',
       );
 
-      for (
-        let epoch = startEpoch;
-        epoch <= epochInfo.getCurrentEpoch();
-        epoch++
-      ) {
+      for (let epoch = startEpoch; epoch <= getCurrentEpoch(); epoch++) {
         const { startCalcTime, endCalcTime } =
           await resolveEpochCalcTimeInterval(epoch);
